feat: mockup db for watcher ids
