+++ .gitignore
@@ -1,2 +1,6 @@
-.idea
+.env
+src
 device.key
+aleph_program.txt
+__pycache__
+.idea
