+++ back/customTypes.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Watcher(BaseModel):
+    address: str
+    network: str
+    watcherType: str
+    confirmationBlocks: str = 1
