+++ back/mockDb.py
@@ -1,0 +1,3 @@
+from customTypes import Watcher
+
+eventMapper = {}
